clean up experiment.py file
