+++ .gitignore
@@ -1,5 +1,6 @@
 .ipynb_checkpoints
-__pycache__
+__pycache__/
+*.pyc
 AllOracleRuns/
 *.out
 manufa_model/
